+++ AzureCommunicationUI/AzureCommunicationUIDemoApp/SceneDelegate.swift
@@ -11,8 +11,12 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
     let envConfigSubject = EnvConfigSubject()
 
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
+    func scene(_ scene: UIScene,
+               willConnectTo session: UISceneSession,
+               options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else {
+            return
+        }
         let entryViewController = EntryViewController(envConfigSubject: envConfigSubject)
         let rootNavController = UINavigationController(rootViewController: entryViewController)
         let window = UIWindow(windowScene: windowScene)
@@ -20,7 +24,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         self.window = window
         window.makeKeyAndVisible()
 
-        //Handle deep link jump from re-launch
+        // Handle deep link jump from re-launch
         let urlContexts = connectionOptions.urlContexts
         if let queryDict = urlContexts.first?.url.toQueryDictionary() {
             envConfigSubject.update(from: queryDict)
@@ -28,7 +32,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     }
 
-    func scene(_ scene: UIScene, openURLContexts URLContexts: Set<UIOpenURLContext>) {
+    func scene(_ scene: UIScene,
+               openURLContexts URLContexts: Set<UIOpenURLContext>) {
 
         if let queryDict = URLContexts.first?.url.toQueryDictionary() {
             envConfigSubject.update(from: queryDict)